Fix dynamic showing context menu
